requirements.txt
black
boto3
botocore
click
colorama
dict2xml
jmespath
mypy-extensions
pathspec
platformdirs
python-dateutil
python-dotenv
s3transfer
six
tomli
urllib3
awswrangler
pandas
paramiko